[refactor] #48 data layer 패키지 구조 정리
