--- app/src/main/java/org/android/bbangzip/data/dto/response/ResponseCompleteTimerDto.kt
@@ -1,16 +1,0 @@
-package org.android.bbangzip.data.dto.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import org.android.bbangzip.domain.model.CompleteBreadCountEntity
-
-@Serializable
-data class ResponseCompleteTimerDto(
-    @SerialName("count")
-    val count: Int,
-) {
-    fun toCompleteTimerEntity() =
-        CompleteBreadCountEntity(
-            count = count,
-        )
-}
