+++ app/src/main/java/org/android/bbangzip/data/repository/di/RepositoryModule.kt
@@ -1,0 +1,29 @@
+package org.android.bbangzip.data.repository.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import org.android.bbangzip.data.repository.local.UserRepositoryImpl
+import org.android.bbangzip.data.repository.remote.DummyRepositoryImpl
+import org.android.bbangzip.data.repository.remote.TimerRepositoryImpl
+import org.android.bbangzip.domain.repository.local.UserRepository
+import org.android.bbangzip.domain.repository.remote.DummyRepository
+import org.android.bbangzip.domain.repository.remote.TimerRepository
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+    @Binds
+    @Singleton
+    abstract fun bindsUserRepository(repositoryImpl: UserRepositoryImpl): UserRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindsDummyRepository(repositoryImpl: DummyRepositoryImpl): DummyRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindsTimerRepository(repositoryImpl: TimerRepositoryImpl): TimerRepository
+}
