+++ app/src/main/java/org/android/bbangzip/data/source/remote/dto/response/ResponseDummyDto.kt
@@ -1,0 +1,16 @@
+package org.android.bbangzip.data.source.remote.dto.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import org.android.bbangzip.domain.model.DummyEntity
+
+@Serializable
+data class ResponseDummyDto(
+    @SerialName("dummy")
+    val dummyName: String,
+) {
+    fun toDummyEntity() =
+        DummyEntity(
+            dummyName = dummyName,
+        )
+}
