+++ app/src/main/java/org/android/bbangzip/data/source/remote/service/DummyService.kt
@@ -1,0 +1,15 @@
+package org.android.bbangzip.data.source.remote.service
+
+import org.android.bbangzip.data.source.remote.dto.response.ResponseDummyDto
+import org.android.bbangzip.data.source.remote.util.base.BaseResponse
+import org.android.bbangzip.data.source.remote.util.constant.ApiConstants.API
+import org.android.bbangzip.data.source.remote.util.constant.ApiConstants.ID
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface DummyService {
+    @GET("$API/{$ID}")
+    suspend fun getDummy(
+        @Path(ID) dummyId: Long,
+    ): BaseResponse<ResponseDummyDto>
+}
