+++ app/src/main/java/org/android/bbangzip/data/source/remote/util/base/BaseResponse.kt
@@ -1,0 +1,14 @@
+package org.android.bbangzip.data.source.remote.util.base
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BaseResponse<T>(
+    @SerialName("code")
+    val code: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: T,
+)
