+++ app/src/main/java/org/android/bbangzip/data/source/remote/util/constant/ApiConstants.kt
@@ -1,0 +1,10 @@
+package org.android.bbangzip.data.source.remote.util.constant
+
+object ApiConstants {
+    const val API = "api"
+    const val ID = "id"
+
+    const val VERSIONS = "v1"
+
+    const val TIMER = "timers"
+}
