+++ app/src/main/java/org/android/bbangzip/data/source/remote/util/qualifier/Qualifier.kt
@@ -1,0 +1,11 @@
+package org.android.bbangzip.data.source.remote.util.qualifier
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class BbangZip
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class Auth
