--- app/src/main/java/org/android/bbangzip/di/KakaoModule.kt
@@ -1,14 +1,0 @@
-package org.android.bbangzip.di
-
-import com.kakao.sdk.user.UserApiClient
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object KakaoModule {
-    @Provides
-    fun providesUserApiClient(): UserApiClient = UserApiClient.instance
-}
