--- app/src/main/java/org/android/bbangzip/di/Qualifier.kt
@@ -1,11 +1,0 @@
-package org.android.bbangzip.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class BbangZip
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class Auth
