--- app/src/main/java/org/android/bbangzip/di/RetrofitModule.kt
@@ -1,76 +1,0 @@
-package org.android.bbangzip.di
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-import okhttp3.Interceptor
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import org.android.bbangzip.BuildConfig
-import org.android.bbangzip.BuildConfig.DEBUG
-import org.android.bbangzip.data.interceptor.AuthInterceptor
-import retrofit2.Retrofit
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object RetrofitModule {
-    @OptIn(ExperimentalSerializationApi::class)
-    @Provides
-    @Singleton
-    fun providesJson(): Json =
-        Json {
-            isLenient = true
-            prettyPrint = true
-            explicitNulls = false
-            ignoreUnknownKeys = true
-        }
-
-    @Provides
-    @Singleton
-    fun providesOkHttpClient(
-        loggingInterceptor: HttpLoggingInterceptor,
-        @Auth authInterceptor: Interceptor,
-    ): OkHttpClient =
-        OkHttpClient.Builder().apply {
-            connectTimeout(10, TimeUnit.SECONDS)
-            writeTimeout(10, TimeUnit.SECONDS)
-            readTimeout(10, TimeUnit.SECONDS)
-            addInterceptor(authInterceptor)
-            if (DEBUG) addInterceptor(loggingInterceptor)
-        }.build()
-
-    @Provides
-    @Singleton
-    fun providesLoggingInterceptor(): HttpLoggingInterceptor =
-        HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        }
-
-    @Provides
-    @Singleton
-    @Auth
-    fun providesAuthInterceptor(interceptor: AuthInterceptor): Interceptor = interceptor
-
-    @ExperimentalSerializationApi
-    @Provides
-    @Singleton
-    @BbangZip
-    fun providesBbangZipRetrofit(
-        okHttpClient: OkHttpClient,
-        json: Json,
-    ): Retrofit =
-        Retrofit.Builder()
-            .baseUrl(BuildConfig.BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(
-                json.asConverterFactory(requireNotNull("application/json".toMediaTypeOrNull())),
-            )
-            .build()
-}
